requirements.txt
matplotlib==3.2.0
aiohttp==3.6.3
discord.py==1.5.1
numpy==1.19.4
discord==1.0.1
python-dotenv==0.15.0
